backend/requirements.txt
PyMuPDF==1.23.6
faiss-cpu==1.7.4
pymongo==4.4.1
sentence-transformers==2.2.2
openai==1.76.2
Flask==2.2.3
Werkzeug==2.2.3
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
Flask-Cors==4.0.0
huggingface_hub==0.15.1